server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
psycopg2-binary
python-dotenv
numpy
Flask-Cors